backend/app/core/supabase.py
from supabase import create_client, Client
import os
from dotenv import load_dotenv
from pathlib import Path

# Load .env from the root directory
env_path = Path(__file__).parent.parent.parent.parent / '.env'
load_dotenv(env_path)

supabase_url = os.getenv("SUPABASE_URL")
supabase_key = os.getenv("SUPABASE_SERVICE_ROLE_KEY")

if not supabase_url or not supabase_key:
    raise ValueError("Missing Supabase credentials in environment variables")

supabase: Client = create_client(supabase_url, supabase_key) 